feat: split lib into modules
